adding remaining files
